requirements.txt
pytest==8.3.4
allure-pytest==2.13.5
selenium==4.27.1
webdriver-manager==4.0.2
faker==33.3.0
requests==2.32.3
